project1/requirements.txt
Django==2.2.8
django-cors-headers==3.2.0
django-oauth2-provider==0.2.6.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.3
shortuuid==0.5.0
sqlparse==0.3.0
uWSGI==2.0.18
